Move constants to own folder
Constants fixes
