+++ kolibri/plugins/coach/assets/src/constants/assignmentsConstants.js
