+++ kolibri/plugins/coach/assets/src/constants/examConstants.js
