+++ kolibri/plugins/coach/assets/src/constants/lessonsConstants.js
