+++ kolibri/plugins/coach/assets/src/state/actions/reports.js
@@ -3,7 +3,7 @@ import { assessmentMetaDataState } from 'kolibri.coreVue.vuex.mappers';
 import * as coreGetters from 'kolibri.coreVue.vuex.getters';
 import * as CoreConstants from 'kolibri.coreVue.vuex.constants';
 import * as Constants from '../../constants';
-import * as ReportConstants from '../../reportConstants';
+import * as ReportConstants from '../../constants/reportConstants';
 import { className } from '../getters/main';
 import { setClassState } from './main';
 import { now } from 'kolibri.utils.serverClock';
