+++ kolibri/plugins/coach/assets/src/views/exams/exam-report-detail-page/index.vue
@@ -69,7 +69,7 @@
 
 <script>
 
-  import * as constants from '../../../constants';
+  import { PageNames } from '../../../constants';
   import immersiveFullScreen from 'kolibri.coreVue.components.immersiveFullScreen';
   import contentRenderer from 'kolibri.coreVue.components.contentRenderer';
   import pageStatus from './page-status';
@@ -88,7 +88,7 @@
     computed: {
       backPageLink() {
         return {
-          name: constants.PageNames.EXAM_REPORT,
+          name: PageNames.EXAM_REPORT,
           params: {
             classId: this.classId,
             channelId: this.channelId,
@@ -106,7 +106,7 @@
       },
       navigateTo(question, interaction) {
         this.$router.push({
-          name: constants.PageNames.EXAM_REPORT_DETAIL,
+          name: PageNames.EXAM_REPORT_DETAIL,
           params: {
             channelId: this.channelId,
             classId: this.classId,
