+++ kolibri/plugins/coach/assets/src/views/reports/item-list-page.vue
@@ -73,7 +73,7 @@
   import { TopicReports, LearnerReports, PageNames } from '../../constants';
   import { ContentNodeKinds } from 'kolibri.coreVue.vuex.constants';
   import { exerciseCount, contentCount, standardDataTable } from '../../state/getters/reports';
-  import { TableColumns } from '../../reportConstants';
+  import { TableColumns } from '../../constants/reportConstants';
   import contentIcon from 'kolibri.coreVue.components.contentIcon';
   import breadcrumbs from './breadcrumbs';
   import headerCell from './table-cells/header-cell';
