+++ kolibri/plugins/coach/assets/src/views/reports/learner-list-page.vue
@@ -77,7 +77,7 @@
   import { ContentNodeKinds } from 'kolibri.coreVue.vuex.constants';
   import { PageNames } from '../../constants';
   import { exerciseCount, contentCount, standardDataTable } from '../../state/getters/reports';
-  import { TableColumns } from '../../reportConstants';
+  import { TableColumns } from '../../constants/reportConstants';
   import contentIcon from 'kolibri.coreVue.components.contentIcon';
   import breadcrumbs from './breadcrumbs';
   import headerCell from './table-cells/header-cell';
