+++ kolibri/plugins/coach/assets/src/views/top-nav.vue
@@ -45,7 +45,7 @@
 <script>
 
   import { PageNames } from '../constants';
-  import { LessonsPageNames } from '../lessonsConstants';
+  import { LessonsPageNames } from '../constants/lessonsConstants';
   import kNavbar from 'kolibri.coreVue.components.kNavbar';
   import kNavbarLink from 'kolibri.coreVue.components.kNavbarLink';
   export default {
